requirements-production.txt
# EduRisk AI - Production Dependencies (Optimized for Docker)
# Backend API Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database & ORM
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Data Validation
pydantic==2.7.4
pydantic-settings==2.3.3

# ML & Data Science
xgboost==2.0.3
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1

# Explainability
shap==0.44.1

# LLM Integration
anthropic==0.28.0

# Caching & Rate Limiting
redis==5.0.4
hiredis==2.3.2

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
python-json-logger==2.0.7
